+++ CHANGELOG.md
@@ -8,11 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Update the README.
-- Improvements for literal and object
-- New type - struct.
 - All types are methods.
 - Now it is possible to get default values. The `real` option has been replaced by the `mode: 'real' | 'current' | 'default'`.
+- Replace arguments with option object: array, literal and object.
+- Add new literal ability: type validator as a choice.
+- Add new type - struct.
 - Add failThrow and config normal methods (e.g. unset, set, load).
+- Add number and integer options: pattern, min and max.
+- Add string options: pattern.
 
 ## [0.2.0] - 2024-10-18